add all clients
