+++ composer/utils/eval_client/eval_client.py
@@ -1,12 +1,13 @@
+# Copyright 2022 MosaicML Composer authors
+# SPDX-License-Identifier: Apache-2.0
+
 # Copyright 2023 MosaicML Composer authors
 # SPDX-License-Identifier: Apache-2.0
 
 """Abstract class for utilities that access and run code on serverless eval clients."""
 
 import abc
-import pathlib
-from types import TracebackType
-from typing import Callable, Optional, Type, Union
+from typing import Dict
 
 __all__ = ['EvalClient']
 
@@ -14,59 +15,26 @@ __all__ = ['EvalClient']
 class EvalClient(abc.ABC):
     """Abstract class for implementing eval clients, such as LambdaEvalClient."""
 
-    def get_uri(self, object_name: str) -> str:
-        """Returns the URI for ``object_name``.
+    def invoke(self, payload: Dict[str, str]) -> bool:
+        """Invoke a provided dictionary payload to the client.
 
-        .. note::
-
-            This function does not check that ``object_name`` is in the object store.
-            It computes the URI statically.
+        For code generation, the payload has format:
+        {
+            'code': <code to be evaluated>,
+            'input': <test input>,
+            'output': <test output>,
+            'entry_point': <entry point>,
+        }
 
         Args:
-            object_name (str): The object name.
+            payload: the materials of the HTTPS request to the client.
 
         Returns:
-            str: The URI for ``object_name`` in the object store.
+            Whether the test case passed or failed.
         """
-        raise NotImplementedError(f'{type(self).__name__}.get_uri is not implemented')
-
-
-    def download_object(
-        self,
-        object_name: str,
-        filename: Union[str, pathlib.Path],
-        overwrite: bool = False,
-        callback: Optional[Callable[[int, int], None]] = None,
-    ) -> None:
-        """Download an object to the specified destination path.
-
-        Args:
-            object_name (str): The name of the object to download.
-            filename (str | pathlib.Path): Full path to a file or a directory where the incoming file will be saved.
-            overwrite (bool, optional): Whether to overwrite an existing file at ``filename``, if it exists.
-                (default: ``False``)
-            callback ((int) -> None, optional): If specified, the callback is periodically called with the number of bytes already
-                downloaded and the total size of the object.
-
-        Raises:
-            FileNotFoundError: If the file was not found in the object store.
-            ObjectStoreTransientError: If there was a transient connection issue with downloading the object.
-        """
-        del object_name, filename, overwrite, callback  # unused
-        raise NotImplementedError(f'{type(self).__name__}.download_object is not implemented')
+        del payload  # unused
+        raise NotImplementedError(f'{type(self).__name__}.invoke is not implemented')
 
     def close(self):
         """Close the object store."""
         pass
-
-    def __enter__(self):
-        return self
-
-    def __exit__(
-        self,
-        exc_type: Optional[Type[BaseException]],
-        exc: Optional[BaseException],
-        traceback: Optional[TracebackType],
-    ):
-        del exc_type, exc, traceback  # unused
-        self.close()
